chore: revert int to int64 conversion

+++ pkg/razorpay/tools_params.go
@@ -102,23 +102,23 @@ func OptionalParam[T any](r mcpgo.CallToolRequest, name string) (T, error) {
 }
 
 // RequiredInt extracts a required integer parameter from the request
-func RequiredInt(r mcpgo.CallToolRequest, name string) (int64, error) {
+func RequiredInt(r mcpgo.CallToolRequest, name string) (int, error) {
 	v, err := RequiredParam[float64](r, name)
 	if err != nil {
 		return 0, err
 	}
 
-	return int64(v), nil
+	return int(v), nil
 }
 
 // OptionalInt extracts an optional integer parameter from the request
-func OptionalInt(r mcpgo.CallToolRequest, name string) (int64, error) {
+func OptionalInt(r mcpgo.CallToolRequest, name string) (int, error) {
 	v, err := OptionalParam[float64](r, name)
 	if err != nil {
 		return 0, err
 	}
 
-	return int64(v), nil
+	return int(v), nil
 }
 
 // AddPaginationToQueryParams processes and adds pagination parameters
